fix ringtone resource permission issues

+++ services/ringtone_scanner/src/ringtone_scanner_db.cpp
@@ -172,17 +172,14 @@ static void InsertDateAdded(const RingtoneMetadata &metadata, ValuesBucket &outV
             int64_t dateModified = metadata.GetDateModified();
             if (dateModified == 0) {
                 dateAdded = RingtoneFileUtils::UTCTimeMilliSeconds();
-                RINGTONE_WARN_LOG("Invalid dateAdded time, use current time instead: %{public}" PRId64,
-                    static_cast<long long>(dateAdded));
+                RINGTONE_WARN_LOG("Invalid dateAdded time, use current time instead: %{public}" PRId64, dateAdded);
             } else {
                 dateAdded = dateModified;
-                RINGTONE_WARN_LOG("Invalid dateAdded time, use dateModified instead: %{public}" PRId64,
-                    static_cast<long long>(dateAdded));
+                RINGTONE_WARN_LOG("Invalid dateAdded time, use dateModified instead: %{public}" PRId64, dateAdded);
             }
         } else {
             dateAdded = dateTaken * MSEC_TO_SEC;
-            RINGTONE_WARN_LOG("Invalid dateAdded time, use dateTaken instead: %{public}" PRId64,
-                static_cast<long long>(dateAdded));
+            RINGTONE_WARN_LOG("Invalid dateAdded time, use dateTaken instead: %{public}" PRId64, dateAdded);
         }
     }
     outValues.PutLong(RINGTONE_COLUMN_DATE_ADDED, dateAdded);
